251031/Tarasenko/lab 5/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
cassandra-driver
aioredis
pydantic
python-dotenv
